requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask_login
psycopg2-binary
pandas
python-dotenv